Fix loading tx from snapshot on dashboard
A referenced input in a tx loaded from the snapshot comes from the
genesis transaction. Genesis transaction is not a real transaction
in a sense that it doesn't have inputs, as this tx creates the funds
available in the network.

Manually fill out info of outputs coming from the genesis tx on
local dashboard. Address "GENESIS" is displayed with 0 balance
for such referenced outputs.