Do not unwrap leaf error into a nil
When dashboard basic auth was enabled and user not authenticated,
the echo server return a HTTP 500 response. The root cause
was that in the error handling, the error was already leaf error,
but it get unwrapped to nil get handled as internal server error.

Use the unwrapAll() which returns the leaf node always, and error
get handled properly.